src/lambda/agentcore-strands/pyproject.toml
[project]
name = "agentcore-strands"
version = "0.1.0"
description = "Slack Bot AgentCore Runtime with Strands Graph"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore[strands-agents]>=0.1.0",
    "strands-agents>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
